+++ packages/platforms/react-native/tests/auto-instrumentation/app-start-plugin.test.ts
@@ -1,15 +1,17 @@
+import type { AppState, Clock, SetAppState } from '@bugsnag/core-performance'
 import { MockSpanFactory, createConfiguration } from '@bugsnag/js-performance-test-utilities'
-import type { Clock } from '@bugsnag/core-performance'
-import createClock from '../../lib/clock'
-import { AppStartPlugin } from '../../lib/auto-instrumentation/app-start-plugin'
-import type { ReactNativeConfiguration } from '../../lib/config'
 import type { AppRegistry } from 'react-native'
+import { AppStartPlugin } from '../../lib/auto-instrumentation/app-start-plugin'
+import createClock from '../../lib/clock'
+import type { ReactNativeConfiguration } from '../../lib/config'
 import type { ReactNativeSpanFactory } from '../../lib/span-factory'
 
 describe('app start plugin', () => {
   let spanFactory: MockSpanFactory<ReactNativeConfiguration>
   let clock: Clock
   let appRegistry: typeof AppRegistry
+  const setAppState: SetAppState = jest.fn()
+  const appState: AppState = 'starting'
 
   beforeEach(() => {
     spanFactory = new MockSpanFactory()
@@ -21,7 +23,7 @@ describe('app start plugin', () => {
 
   it('starts an app start span when autoInstrumentAppStarts is true', () => {
     const appStartTime = 1234
-    const plugin = new AppStartPlugin(appStartTime, spanFactory as unknown as ReactNativeSpanFactory, clock, appRegistry)
+    const plugin = new AppStartPlugin(appStartTime, spanFactory as unknown as ReactNativeSpanFactory, clock, appRegistry, setAppState, appState)
 
     plugin.configure(createConfiguration<ReactNativeConfiguration>({ autoInstrumentAppStarts: true }))
 
@@ -35,7 +37,7 @@ describe('app start plugin', () => {
   })
 
   it('does not start an app start span when autoInstrumentAppStarts is false', () => {
-    const plugin = new AppStartPlugin(1234, spanFactory as unknown as ReactNativeSpanFactory, clock, appRegistry)
+    const plugin = new AppStartPlugin(1234, spanFactory as unknown as ReactNativeSpanFactory, clock, appRegistry, setAppState, appState)
 
     plugin.configure(createConfiguration<ReactNativeConfiguration>({ autoInstrumentAppStarts: false }))