+++ packages/plugin-react-native-navigation/lib/react-native-navigation-plugin.ts
@@ -1,4 +1,4 @@
-import type { Plugin, SpanFactory, SpanInternal } from '@bugsnag/core-performance'
+import type { Plugin, SetAppState, SpanFactory, SpanInternal } from '@bugsnag/core-performance'
 import type { ReactNativeConfiguration, ReactNativeSpanFactory } from '@bugsnag/react-native-performance'
 import type { NavigationDelegate } from 'react-native-navigation/lib/dist/src/NavigationDelegate'
 
@@ -16,6 +16,7 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
   private componentsWaiting = 0
   private spanFactory?: ReactNativeSpanFactory
   private previousRoute?: string
+  private setAppState?: SetAppState
 
   constructor (Navigation: NavigationDelegate) {
     this.Navigation = Navigation
@@ -33,6 +34,9 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
       this.currentNavigationSpan.setAttribute('bugsnag.navigation.ended_by', endedBy)
       this.spanFactory.endSpan(this.currentNavigationSpan, endTime)
       this.clearActiveSpan()
+      if (this.setAppState) {
+        this.setAppState('ready')
+      }
     }
   }
 
@@ -64,8 +68,9 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
     }
   }
 
-  configure (configuration: ReactNativeConfiguration, spanFactory: SpanFactory<ReactNativeConfiguration>) {
+  configure (_configuration: ReactNativeConfiguration, spanFactory: SpanFactory<ReactNativeConfiguration>, setAppState: SetAppState) {
     this.spanFactory = spanFactory as ReactNativeSpanFactory
+    this.setAppState = setAppState
 
     // Potential for a navigation to occur
     this.Navigation.events().registerCommandListener((name, params) => {
@@ -79,8 +84,11 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
     // Navigation has occurred
     this.Navigation.events().registerComponentWillAppearListener(event => {
       if (this.spanFactory && typeof this.startTime === 'number') {
-        clearTimeout(this.startTimeout)
+        if (this.setAppState) {
+          this.setAppState('navigating')
+        }
 
+        clearTimeout(this.startTimeout)
         const routeName = event.componentName
         this.currentNavigationSpan = this.spanFactory.startNavigationSpan(routeName, { startTime: this.startTime, doNotDelegateToNativeSDK: true })
         this.currentNavigationSpan.setAttribute('bugsnag.navigation.triggered_by', '@bugsnag/plugin-react-native-navigation-performance')