+++ packages/plugin-react-navigation/test/create-navigation-container.test.tsx
@@ -1,10 +1,11 @@
+import type { AppState } from '@bugsnag/core-performance'
 import { MockReactNativeSpanFactory } from '@bugsnag/js-performance-test-utilities'
-import { NavigationContainer, createNavigationContainerRef } from '@react-navigation/native'
+import type { ReactNativeSpanFactory } from '@bugsnag/react-native-performance'
 import type { ParamListBase } from '@react-navigation/native'
+import { NavigationContainer, createNavigationContainerRef } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import '@testing-library/jest-native/extend-expect'
 import { act, fireEvent, render, screen } from '@testing-library/react-native'
-import React from 'react'
 import { Button, Text, View } from 'react-native'
 import { createNavigationContainer } from '../lib/create-navigation-container'
 
@@ -18,8 +19,12 @@ afterEach(() => {
 
 describe('createNavigationContainer', () => {
   it('creates a navigation span when the route changes', () => {
+    let appState: AppState = 'starting'
+    const setAppState = jest.fn((state: AppState) => {
+      appState = state
+    })
     const spanFactory = new MockReactNativeSpanFactory()
-    const BugsnagNavigationContainer = createNavigationContainer(NavigationContainer, spanFactory)
+    const BugsnagNavigationContainer = createNavigationContainer(NavigationContainer, spanFactory as unknown as ReactNativeSpanFactory, setAppState)
 
     render(
       <BugsnagNavigationContainer>
@@ -32,18 +37,25 @@ describe('createNavigationContainer', () => {
 
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledTimes(1)
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledWith('Route 2', { isFirstClass: true, startTime: 0, doNotDelegateToNativeSDK: true })
+    expect(setAppState).toHaveBeenCalled()
+    expect(appState).toBe('navigating')
 
     fireEvent.press(screen.getByText('Go back'))
     expect(screen.getByText('Route 1')).toBeOnTheScreen()
 
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledTimes(2)
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledWith('Route 1', { isFirstClass: true, startTime: 0, doNotDelegateToNativeSDK: true })
+    expect(setAppState).toHaveBeenCalledTimes(2)
   })
 
   it('forwards the provided ref to the NavigationContainer', () => {
+    let appState: AppState = 'starting'
+    const setAppState = jest.fn((state: AppState) => {
+      appState = state
+    })
     const navigationRef = createNavigationContainerRef()
     const spanFactory = new MockReactNativeSpanFactory()
-    const BugsnagNavigationContainer = createNavigationContainer(NavigationContainer, spanFactory)
+    const BugsnagNavigationContainer = createNavigationContainer(NavigationContainer, spanFactory as unknown as ReactNativeSpanFactory, setAppState)
 
     render(
       <BugsnagNavigationContainer ref={navigationRef}>
@@ -58,6 +70,8 @@ describe('createNavigationContainer', () => {
 
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledTimes(1)
     expect(spanFactory.startNavigationSpan).toHaveBeenCalledWith('Route 2', { isFirstClass: true, startTime: 0, doNotDelegateToNativeSDK: true })
+    expect(setAppState).toHaveBeenCalledTimes(1)
+    expect(appState).toBe('navigating')
   })
 })