+++ test/browser/features/fixtures/packages/react-router/rollup.config.mjs
@@ -24,6 +24,7 @@ const cdnConfig = {
       'react-dom/client': 'ReactDom',
       'react-router-dom': 'ReactRouterDom',
       '@bugsnag/react-router-performance': 'BugsnagReactRouterPerformance',
+      '@bugsnag/plugin-react-performance': 'BugsnagPluginReactPerformance',
     },
   }
 }
@@ -33,8 +34,12 @@ export default {
   input: 'src/index.jsx',
   plugins: [
     nodeResolve({ browser: true, jail: path.resolve(`${__dirname}/../..`), extensions: ['.mjs', '.js', '.json', '.node', '.jsx'] }),
+    babel({ 
+      babelHelpers: 'bundled',
+      presets: ['@babel/preset-react'],
+      extensions: ['.js', '.jsx']
+    }),
     commonjs(),
-    babel({ babelHelpers: 'bundled' }),
     replace({
       preventAssignment: true,
       values: {